fix(chartcards): add truncation prop and enablement
